[#629] AI 키보드 입력 Scene을 CalendarScenes → AIAgentScene 이동
AI 자연어 입력 관련 컴포넌트를 AIAgentScene 프레임워크에 응집.

- AIAgentKeyboardInput 5파일(Builder/VC/VM/Router/View)을 AIAgentScene로 이동.
- Scenes+AIAgent — AIAgentKeyboardInputScene/Interactor/SceneBuilder 공유 프로토콜 public 노출(makeScene → makeKeyboardInputScene). CalendarScenes 로컬 프로토콜 제거.
- AIAgentKeyboardInputBuilderImple — aiAgentOrchestrationUsecase 직접 주입 → usecaseFactory 주입으로 전환(command builder와 동일 패턴). 공유 orchestration usecase 인스턴스라 DayEventList 상태 공유 유지. public화.
- 배선 — ApplicationRootRouter가 keyboard builder 생성 → CalendarSceneBuilderImple → DayEventListSceneBuilerImple 주입. CalendarScenes는 AIAgentScene 직접 import 없이 Scenes 공유 프로토콜로만 참조(presentations-rules §7).
- AIAgentKeyboardInputView — backgroundAsRoundedRectForEventList(CalendarScenes 전용 헬퍼) 인라인화 + 프리뷰 추가.
