+++ Presentations/AIAgentScene/Sources/AIAgentKeyboardInputViewModel.swift
@@ -1,6 +1,6 @@
 //
 //  AIAgentKeyboardInputViewModel.swift
-//  CalendarScenes
+//  AIAgentScene
 //
 
 import Foundation
