--- Presentations/CalendarScenes/Sources/DayEventList/AIAgentKeyboardInputScene+Builder.swift
@@ -1,25 +1,0 @@
-//
-//  AIAgentKeyboardInputScene+Builder.swift
-//  CalendarScenes
-//
-
-import UIKit
-import Scenes
-
-
-// MARK: - AIAgentKeyboardInputScene Interactable & Listenable
-
-protocol AIAgentKeyboardInputSceneInteractor: AnyObject { }
-
-// MARK: - AIAgentKeyboardInputScene
-
-protocol AIAgentKeyboardInputScene: Scene where Interactor == any AIAgentKeyboardInputSceneInteractor { }
-
-
-// MARK: - Builder
-
-protocol AIAgentKeyboardInputSceneBuilder: AnyObject {
-
-    @MainActor
-    func makeScene() -> any AIAgentKeyboardInputScene
-}
